Integrate apply for leave API

+++ lib/model/leave_request_data.dart
@@ -3,25 +3,25 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 part 'leave_request_data.g.dart';
 
 @JsonSerializable()
-class LeaveRequestData{
+class LeaveRequestData {
   @JsonKey(name: 'start_date')
   int startDate;
-  @JsonKey(name:'end_date')
+  @JsonKey(name: 'end_date')
   int endDate;
   @JsonKey(name: 'total_leaves')
-  int totalLeaves;
+  double totalLeaves;
   String reason;
   @JsonKey(name: 'emergency_contact_person')
   int emergencyContactPerson;
 
-  LeaveRequestData({
-   required this.startDate,
-    required this.endDate,
-    required this.totalLeaves,
-    required this.reason,
-    required this.emergencyContactPerson
-});
+  LeaveRequestData(
+      {required this.startDate,
+      required this.endDate,
+      required this.totalLeaves,
+      required this.reason,
+      required this.emergencyContactPerson});
 
-  Map<String,dynamic> leaveRequestDataToJson(LeaveRequestData leaveRequestData)=> _$LeaveRequestDataToJson(this);
-
+  Map<String, dynamic> leaveRequestDataToJson(
+          LeaveRequestData leaveRequestData) =>
+      _$LeaveRequestDataToJson(this);
 }
